Added Ctrl, Shift, Alt to Layer 2 for Ctrl+Alt+F<n>. Tidied keymap.
Added the Ctrl Qukeys to Layer 2, as well as Alt, so that I can press
Ctrl+Alt+Function Key to switch VTs. Decided to move all the Qukey
assignments to the QUKEYS() in setup().

Added the Shift Qukeys also to Layer 2 at the same time for good measure.

Tidied up the keymap so it looks pretty and all the layers are aligned.